+++ contrib/net/http/http_test.go
@@ -6,14 +6,6 @@
 package http
 
 import (
-	"io"
-	"net/http"
-	"net/http/httptest"
-	"strings"
-	"sync"
-	"testing"
-	"time"
-
 	"gopkg.in/DataDog/dd-trace-go.v1/contrib/internal/namingschematest"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/ext"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/mocktracer"
@@ -21,9 +13,13 @@ import (
 	"gopkg.in/DataDog/dd-trace-go.v1/internal/globalconfig"
 	"gopkg.in/DataDog/dd-trace-go.v1/internal/log"
 	"gopkg.in/DataDog/dd-trace-go.v1/internal/normalizer"
+	"net/http"
+	"net/http/httptest"
+	"strings"
+	"sync"
+	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestWithHeaderTags(t *testing.T) {
@@ -524,48 +520,6 @@ func TestServerNamingSchema(t *testing.T) {
 		return mt.FinishedSpans()
 	})
 	namingschematest.NewHTTPServerTest(genSpans, "http.router")(t)
-}
-
-// TestUnwrap tests the implementation of the rwUnwrapper interface, which is used internally
-// by the standard library: https://github.com/golang/go/blob/6d89b38ed86e0bfa0ddaba08dc4071e6bb300eea/src/net/http/responsecontroller.go#L42-L44
-// See also: https://github.com/DataDog/dd-trace-go/issues/2674
-func TestUnwrap(t *testing.T) {
-	fn := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		rc := http.NewResponseController(w)
-
-		// Use the SetReadDeadline and SetWriteDeadline methods, which are not explicitly implemented in the trace_gen.go
-		// generated file. Before the Unwrap change, these methods returned a "feature not supported" error.
-
-		err := rc.SetReadDeadline(time.Now().Add(10 * time.Second))
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			w.Write([]byte(err.Error()))
-			return
-		}
-		err = rc.SetWriteDeadline(time.Now().Add(10 * time.Second))
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			w.Write([]byte(err.Error()))
-			return
-		}
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte("OK"))
-	})
-
-	h := WrapHandler(fn, "service-name", "resource-name")
-	srv := httptest.NewServer(h)
-	defer srv.Close()
-
-	resp, err := srv.Client().Get(srv.URL + "/")
-	require.NoError(t, err)
-	defer resp.Body.Close()
-
-	b, err := io.ReadAll(resp.Body)
-	require.NoError(t, err)
-	respText := string(b)
-
-	assert.Equal(t, http.StatusOK, resp.StatusCode)
-	assert.Equal(t, "OK", respText)
 }
 
 func router(muxOpts ...Option) http.Handler {
